Simplify the 2 methods that gets the animation arity
Change the return type of the 2 methods below to only return the arity
(num animation channels). It is enough to check if the arity is
> 0 to know that the data type is animatable. This makes the methods
usage much cleaner.

GetDatatypeAnimatableAndArity()
GetVariantAnimatableAndArity()

You much also pull the commit with the same name in the qt3dstudio repo
for the Editor to compile after this is pulled.

Change-Id: I3e80d49aa6ddf893be43a359ddaf583f50df9180
Reviewed-by: Antti Määttä <antti.maatta@qt.io>