Add transparency mode selection to standard materials
ASTC textures would need to be decompressed upon loading to scan for
transparency, or the texture files would need their own user-defined
settings to enable transparency. Instead, as a work-around, materials
can be forced to be treated as transparent without manipulating opacity.

This is implemented with a transparency mode drop down list containing
"Default", "Force Transparent" and "Force Opaque" options. "Force
Opaque" can be used to force the material to be treated as opaque even
when the textures used are transparent.

By default ASTC textures are treated as non-transparent.

Task-number: QT3DS-3904
Change-Id: I22255ffb2e2be5efea3e56466d58ddcc46490bb2
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Janne Kangas <janne.kangas@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>