Fix subpresentation transparent clear
Always include alpha channel for subpresentations and properly handle
clear color setup. It needs to setup differently when we render to texture
vs viewport. When rendering to viewport, if the clear color contains
alpha, we need to blend the clear color to the target instead of clearing
the target. Also we need to blend the scene color to the matte color, when
matte and scene color are enabled and the scene color contains alpha.

Task-number: QT3DS-4008
Change-Id: Ic2b3bac8ea586cbd205421c8efd21c1e3e49abd4
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>