Fix distance field text alignment
Distance field bounding boxes were working wrong compared to the old
text rendering. When text alignment is set to right, the bounding box
has to be set so the right side is in the center of the object.
Previously the bounding box remained the same, giving an illusion that
the text just shifted to the left, when it really just aligned to the
right of center. This should make eliding work correctly too.

Task-number: QT3DS-4020
Change-Id: I5c946c07f49eb4ca4ce63b45f897def41053ae24
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>