Load referenced material assets upon use
When using delayed loading, assets used by materials inside the
material container are not loaded upon startup anymore. Instead
the assets are loaded when entering the slide in which the material
is referenced.

This also fixes referenced materials that refer to an animatable
material in another slide.

Task-number: QT3DS-3100
Change-Id: If55c73e8f91389f3c47ae16357ebe6a473f43585
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Janne Kangas <janne.kangas@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>