Add dynamic layer resizing
Add a boolean property to layer that enables dynamic resizing. The
screen space bounds of the layer content are calculated and the layer
resized to fit the content and exclude empty space.

Padding property is also added that pads the resized layer so that
minimal changes to the content won't trigger another resize. Padding can
be set by pixel amount or by percentage of the size.

To provide accurate 2d bounds, each active objects bounds are projected
separately to the screen. A boolean property to only project combined
bounds is provided but lowers the accuracy while limiting projections to
8.

Known issues:
 - Temporal AA doesn't work with resizing correctly
 - Temporal AA leaks GPU memory when resizing
 - Editor picking doesn't work correctly with dynamically resized layers
 - Render to viewport doesn't yet support resizing

Task-number: QT3DS-4030
Change-Id: Ibc727e078129eee15021cc97105b4cdf9a5a3196
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>